Change some types for storage of left and right of in and out separately
Only declarations modified, code not always corrected to suit

git-svn-id: svn+ssh://svn.forge.ecs.soton.ac.uk/projects/sanchp3yp@106 bc9da1e1-76ae-0410-a611-e94dc21032b9

+++ work/code/ring.h
@@ -5,14 +5,14 @@
 #include <tistdtypes.h>
 
 typedef struct {
-	int32_t * ring;
-	int32_t size;
+	int16_t * ring;
+	int16_t size;
 } ringbuf;
 
-void init_ring(void *ring, int32_t * arr, int size);
-void init_ring_ptr(void *ring, int32_t ** ptr);
+void init_ring(void *ring, int16_t * arr, int size);
+void init_ring_ptr(void *ring, int16_t ** ptr);
 
-void inc_ring(void * ring, int32_t ** ptr);
-void dec_ring(void * ring, int32_t ** ptr);
+void inc_ring(void * ring, int16_t ** ptr);
+void dec_ring(void * ring, int16_t ** ptr);
 
 #endif /*RING_H_*/
